move push and replace to types
