+++ src/utilities/createRouter.ts
@@ -4,14 +4,14 @@ import { routerInjectionKey, routerRejectionKey } from '@/compositions'
 import { NavigationAbortError } from '@/errors/navigationAbortError'
 import { Routes, Router, RouterOptions, RouterImplementation } from '@/types'
 import { RouterPushError, RouterRejectionError } from '@/types/errors'
+import { RouterPushImplementation } from '@/types/routerPush'
+import { RouterReplaceImplementation } from '@/types/routerReplace'
 import { createCurrentRoute } from '@/utilities/createCurrentRoute'
 import { createRouteMethods } from '@/utilities/createRouteMethods'
 import { createRouterFind } from '@/utilities/createRouterFind'
 import { createRouterHistory } from '@/utilities/createRouterHistory'
 import { addRouteHookInjectionKey, createRouterHooks } from '@/utilities/createRouterHooks'
-import { RouterPushImplementation } from '@/utilities/createRouterPush'
 import { createRouterReject } from '@/utilities/createRouterReject'
-import { RouterReplaceImplementation } from '@/utilities/createRouterReplace'
 import { createRouterResolve } from '@/utilities/createRouterResolve'
 import { createRouterRoutes } from '@/utilities/createRouterRoutes'
 import { getInitialUrl } from '@/utilities/getInitialUrl'
