+++ src/utilities/index.ts
@@ -6,9 +6,9 @@ export * from './createRouteMethods'
 export * from './createRouter'
 export * from './createRouterFind'
 export * from './createRouterHooks'
-export * from './createRouterPush'
+export * from '../types/routerPush'
 export * from './createRouterReject'
-export * from './createRouterReplace'
+export * from '../types/routerReplace'
 export * from './createRouterResolve'
 export * from './createRouterRoute'
 export * from './createRouterRoutes'
